[65][Refactor] Refactor settings. Split into components
