+++ src/ui/menu-button/menu-button.styles.tsx
@@ -1,0 +1,12 @@
+import { styled } from "@mui/material";
+
+export const Wrapper = styled('div')({
+  marginBottom: '20px',
+});
+
+export const Content = styled('div')({
+  display: 'flex',
+  flexDirection: 'column',
+  width: '40vh',
+  height: '400px',
+});
