+++ src/ui/settings/settings.styles.ts
@@ -1,16 +1,5 @@
 import { styled } from '@mui/material';
 
-export const Wrapper = styled('div')({
-  marginBottom: '20px',
-});
-
-export const Content = styled('div')({
-  display: 'flex',
-  flexDirection: 'column',
-  width: '40vh',
-  height: '400px',
-});
-
 export const Container = styled('div')({
   display: 'flex',
   width: '10%',
